// src/app/core/enums/manual.translate-keys.enum.ts
export enum ManualTranslateKeys {
  Title = 'INFO.MANUAL.TITLE',
  Description = 'INFO.MANUAL.DESCRIPTION',
  Interface = 'INFO.INTERFACE.TITLE',
  MainMap = 'INFO.INTERFACE.MAIN_MAP',
  MainMapDescription = 'INFO.INTERFACE.MAIN_MAP_DESCRIPTION',
  StartPoint = 'INFO.INTERFACE.START_POINT',
  StartPointDescription = 'INFO.INTERFACE.START_POINT_DESCRIPTION',
  FinishPoint = 'INFO.INTERFACE.FINISH_POINT',
  FinishPointDescription = 'INFO.INTERFACE.FINISH_POINT_DESCRIPTION',
  FloorSwitch = 'INFO.INTERFACE.FLOOR_SWITCH',
  FloorSwitchDescription= 'INFO.INTERFACE.FLOOR_SWITCH_DESCRIPTION',
  ZoomEdit = 'INFO.INTERFACE.ZOOM_EDIT',
  ZoomEditDescription = 'INFO.INTERFACE.ZOOM_EDIT_DESCRIPTION',
  Settings = 'INFO.INTERFACE.SETTINGS',
  SettingsDescription = 'INFO.INTERFACE.SETTINGS_DESCRIPTION',
  Manual = 'INFO.INTERFACE.MANUAL',
  ManualDescription = 'INFO.INTERFACE.MANUAL_DESCRIPTION',
  Authors = 'INFO.AUTHORS.TITLE',
  Support = 'INFO.SUPPORT.TITLE',
  Footer = 'INFO.FOOTER',
}